+++ client/src/ui-components/ManagerStatsComponent/MonthlyReports.js
@@ -20,6 +20,7 @@ import {
   Label,
 } from "recharts";
 import { arrowDownIcon, arrowUpIcon } from "../icons";
+import { CustomTooltip } from "./CustomTooltip";
 
 function MonthlyReports(props) {
   const this_month_perc = (
@@ -108,29 +109,6 @@ function UnretrievedMonthlyBar(props) {
     data.push(value);
     return entry;
   });
-  const CustomTooltip = ({ active, payload, label }) => {
-    if (active && payload && payload.length) {
-      let year_info = 2021;
-      if (label === "JAN") year_info = 2022;
-      return (
-        <div className="custom-tooltip">
-          <p className="label-tooltip">{`month : ${label}-${year_info}`}</p>
-          <p className="unretrieved-info">{`unretrieved orders : ${payload[0].value}`}</p>
-          <p className="total-info">{`total orders : ${payload[1].value}`}</p>
-          {payload[0].value && payload[1].value ? (
-            <p>{`% of unretrieved orders : ${(
-              (payload[0].value / payload[1].value) *
-              100
-            ).toFixed(2)}`}</p>
-          ) : (
-            <p>{`% of unretrieved orders : N/A`}</p>
-          )}
-        </div>
-      );
-    }
-
-    return null;
-  };
 
   return (
     <>
