+++ client/src/ui-components/ManagerStatsComponent/WeeklyReports.js
@@ -13,6 +13,7 @@ import {
   Label,
 } from "recharts";
 import { arrowDownIcon, arrowUpIcon } from "../icons";
+import { CustomTooltip } from "./CustomTooltip";
 
 function WeeklyReports(props) {
   const this_week_perc = (
@@ -101,29 +102,6 @@ function UnretrievedWeeklyBar(props) {
     data.push(value);
     return entry;
   });
-
-  const CustomTooltip = ({ active, payload, label }) => {
-    if (active && payload && payload.length) {
-      let year_info = 2021;
-      if (label === 1 || label === 2) year_info = 2022;
-      return (
-        <div className="custom-tooltip">
-          <p className="label-tooltip">{`week : ${label}/${year_info}`}</p>
-          <p className="unretrieved-info">{`unretrieved orders : ${payload[0].value}`}</p>
-          <p className="total-info">{`total orders : ${payload[1].value}`}</p>
-          {payload[0].value && payload[1].value ? (
-            <p>{`% of unretrieved orders : ${(
-              (payload[0].value / payload[1].value) *
-              100
-            ).toFixed(2)}`}</p>
-          ) : (
-            <p>{`% of unretrieved orders : N/A`}</p>
-          )}
-        </div>
-      );
-    }
-    return null;
-  };
 
   return (
     <>
